requirements.txt
langchain>=0.2.0
google-generativeai>=0.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
